Improve Temporal specification compliance and timezone/DST disambiguation
- Fix timezone and DST disambiguation logic in interpretISODateTimeOffset and getEpochNanosecondsFor.
- Query both standard/daylight and former/latter ICU interpretations for local wall time offsets to resolve DST overlap/gap edge cases.
- Fix hoursInDay to return double to support fractional hours.
- Respect parsedOffsetNanoseconds when resolving relativeTo options in TemporalDuration.
- Ensure ZonedDateTime constructor uses direct instant-offset resolution instead of incorrect wall-time translation.

+++ src/runtime/TemporalZonedDateTimeObject.h
@@ -173,7 +173,7 @@ public:
     bool equals(ExecutionState& state, Value otherInput);
 
     // https://tc39.es/proposal-temporal/#sec-get-temporal.zoneddatetime.prototype.hoursinday
-    int hoursInDay(ExecutionState& state);
+    double hoursInDay(ExecutionState& state);
 
     // https://tc39.es/proposal-temporal/#sec-temporal.zoneddatetime.prototype.toinstant
     TemporalInstantObject* toInstant(ExecutionState& state);
